usage-*.sql: remove version-specific comment
This removes the need to delete and re-add the whole table, just to make the
resulting schemas identical. For the sake of the unit tests, I'm
retroactively changing the v1 schema to match. Anyone using the upgrader on a
real usagedb will see the wrong comment in the new database, oh well.