database.py: hush pyflakes on py3
I was trying to be more specific with the exception, by looking for
FileNotFoundError (which is only available on py3). The fallback approach I
was using made pyflakes unhappy, as it thought I was referencing an undefined
variable (which I was, but only for the purpose of discovering that it was
undefined). Looking for the less-specific EnvironmentError is plenty for this
purpose.