change default port= to listen on both IPv4+IPv6
The default was "tcp:4000", which happens to expand into
"tcp:4000:interface=127.0.0.1", which limits the listening socket to IPv4
connections only.

Changing it to "tcp:4000:interface=\:\:1" means "listen on ::1", which is the
IPv6 name for the loopback interface, and gets us a socket that accepts both
IPv4 and IPv6 connections.

Note: this might cause incompatibilities with IPv4-only hosts, if they don't
accept the "::1" name. For these systems, run the server as `twist
wormhole-mailbox --port=tcp:4000" to revert to the old behavior.

closes #16