Replace deprecated pkg_resources.resource_string call
setuptool will drop it in several months. When the tests run, a warning is displayed:
database.py:5: UserWarning: pkg_resources is deprecated as an API. See https://setuptools.pypa.io/en/latest/pkg_resources.html. The pkg_resources package is slated for removal as early as 2025-11-30. Refrain from using this package or pin to Setuptools<81.
  from pkg_resources import resource_string

As the resource_string() call is done in the code, I guess the problem will also occurs for users.

The patch is based on an equivalent backport:
https://importlib-resources.readthedocs.io/en/latest/migration.html#pkg-resources-resource-string